+++ src/index.ts
@@ -2,3 +2,4 @@ export * from "./inline-edit.js";
 export * from "./state.js";
 export * from "./theme.js";
 export * from "./trigger.js";
+export * from "./commands.js";
